pyproject.toml
[project]
name = "patent-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.0.0",
    "anthropic>=0.84.0",
    "pydantic>=2.0.0",
"python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
